requirements.txt
blinker==1.8.2
click==8.1.8
dlib==19.24.8
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.2
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
numpy==1.24.4
opencv-python==4.9.0.80
Pillow==10.1.0
werkzeug==3.0.6
zipp==3.20.2
